requirements.txt
APScheduler==3.6.3
certifi==2024.8.30
python-dotenv==1.0.1
python-telegram-bot==13.5
pytz==2024.2
six==1.17.0
tornado==6.4.2
tzlocal==5.2
